+++ packages/manager/src/features/linodes/LinodesDetail/LinodeStorage/LinodeStorage.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import LinodeDisks from '../LinodeAdvanced/LinodeDisks_CMR';
+import LinodeDisks from '../LinodeAdvanced/LinodeDisks';
 import LinodeVolumes from 'src/features/linodes/LinodesDetail/LinodeAdvanced/LinodeVolumes';
 
 export const LinodeStorage: React.FC = () => {
